libc: arm64: Use SIMD-accelerated memcpy
It turns out bionic is not using SIMD variant of memcpy although NEON can
be used in userspace. Import memcpy-advsimd from optimized-routines and
use it. Guard was added for extra safety but it is probably superfluous
on every modern Android device.

Initial commit in arm repo stated performance increase of 11%.

Change-Id: I6614329221b8aafa54fe3f18b53421bea67a559c
Signed-off-by: Andrzej Perczak <linux@andrzejperczak.com>
Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>